% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_model.R
\name{bin_model}
\alias{bin_model}
\title{bin_model: Binarize a model using a threshold}
\usage{
bin_model(model, occs, percent)
}
\arguments{
\item{model}{A continuous raster model of suitability values}

\item{occs}{Occurrence data with two columns (longitude and latitude).}

\item{percent}{Type of thresholding method. Values go from 0-100 percent
of the data. 0 is equivalent to the minimum training presence.}
}
\value{
A binary map of the prediction.
}
\description{
Function to binarize a continuous model given a threshold
}
\examples{
\dontrun{

model_p <- system.file("extdata/rasters/ambystoma_model.tif",
                       package = "ntbbox")
model <- raster::raster(model_p)
data_p <- system.file("extdata/csv/ambystoma_validation.csv",
                      package = "ntbbox")
data <- read.csv(data_p)
occs <- data[which(data$presence_absence==1),]
binary <- bin_model(model,occs[,1:2],percent = 10)
raster::plot(binary)
}
# Example with data of the giant hummingbird
r1_path <- system.file("extdata/rasters/p_gigas.tif",
                       package = "ntbbox")
r1 <- raster::raster(r1_path)
pgig_p <- system.file("extdata/csv/p_gigas.csv",
                      package = "ntbbox")
pgig <- read.csv(pgig_p)
pgig <- pgig[pgig$type=="train",]
bin_pg <- bin_model(r1,pgig[,2:3],percent = 15)
raster::plot(bin_pg)
}
